refactor(eloquent): extract ModelSerializationShapeBuilder #923
ModelToArrayShapeHandler had grown into a partial reimplementation of
Laravel's attributesToArray() serialization: the visible/hidden filter,
column cast serialization (date/enum), and append serialization
(Arrayable/date/generics) all lived in the provider.

Move that logic into a new pure ModelSerializationShapeBuilder that owns
the serialization order and per-attribute serialized-type mapping,
consuming the warmed ModelMetadata. The handler is now a thin Psalm
event adapter: method-name gate, framework-serializer override bail,
registry lookup, then delegate. Mirrors the ColumnTypeMapper split (a
pure mapper consumed by handlers, autoloaded, no require_once).

Behavior-preserving: no shape output changes. The unit test moves to
ModelSerializationShapeBuilderTest and drives the builder directly; the
type and integration suites are unchanged.
